+++ quacc/recipes/gaussian/core.py
@@ -8,7 +8,7 @@ import covalent as ct
 from ase import Atoms
 from ase.calculators.gaussian import Gaussian
 
-from quacc.schemas.cclib import summarize_run
+from quacc.schemas.cclib import cclibSchema, summarize_run
 from quacc.util.calc import run_calc
 from quacc.util.dicts import remove_dict_empties
 
@@ -18,13 +18,13 @@ GEOM_FILE = LOG_FILE
 
 @ct.electron
 def static_job(
-    atoms: Atoms | dict[Literal["atoms"], Atoms],
+    atoms: Atoms | dict,
     charge: int | None = None,
     multiplicity: int | None = None,
     xc: str = "wb97x-d",
     basis: str = "def2-tzvp",
     calc_swaps: dict | None = None,
-) -> dict:
+) -> cclibSchema:
     """
     Carry out a single-point calculation.
 
@@ -63,7 +63,7 @@ def static_job(
 
     Returns
     -------
-    dict
+    RunSchema
         Dictionary of results from `quacc.schemas.cclib.summarize_run`
     """
     atoms = atoms if isinstance(atoms, Atoms) else atoms["atoms"]
@@ -113,7 +113,7 @@ def relax_job(
     basis: str = "def2-tzvp",
     freq: bool = False,
     calc_swaps: dict | None = None,
-) -> dict:
+) -> cclibSchema:
     """
     Carry out a geometry optimization.
 
@@ -154,7 +154,7 @@ def relax_job(
 
     Returns
     -------
-    dict
+    RunSchema
         Dictionary of results from `quacc.schemas.cclib.summarize_run`
     """
     atoms = atoms if isinstance(atoms, Atoms) else atoms["atoms"]