+++ quacc/recipes/orca/core.py
@@ -8,7 +8,7 @@ import covalent as ct
 from ase import Atoms
 from ase.calculators.orca import ORCA
 
-from quacc.schemas.cclib import summarize_run
+from quacc.schemas.cclib import cclibSchema, summarize_run
 from quacc.util.calc import run_calc
 from quacc.util.dicts import remove_dict_empties
 
@@ -18,14 +18,14 @@ GEOM_FILE = f"{ORCA().name}.xyz"
 
 @ct.electron
 def static_job(
-    atoms: Atoms | dict[Literal["atoms"], Atoms],
+    atoms: Atoms | dict,
     charge: int | None = None,
     multiplicity: int | None = None,
     xc: str = "wb97x-d3bj",
     basis: str = "def2-tzvp",
     input_swaps: dict | None = None,
     block_swaps: dict | None = None,
-) -> dict:
+) -> cclibSchema:
     """
     Carry out a single-point calculation.
 
@@ -63,7 +63,7 @@ def static_job(
 
     Returns
     -------
-    dict
+    cclibSchema
         Dictionary of results from quacc.schemas.cclib.summarize_run
     """
     atoms = atoms if isinstance(atoms, Atoms) else atoms["atoms"]
@@ -113,7 +113,7 @@ def static_job(
 
 @ct.electron
 def relax_job(
-    atoms: Atoms | dict[Literal["atoms"], Atoms],
+    atoms: Atoms | dict,
     charge: int = None,
     multiplicity: int = None,
     xc: str = "wb97x-d3bj",
@@ -121,7 +121,7 @@ def relax_job(
     run_freq: bool = False,
     input_swaps: dict = None,
     block_swaps: dict = None,
-) -> dict:
+) -> cclibSchema:
     """
     Carry out a geometry optimization.
 
@@ -162,7 +162,7 @@ def relax_job(
 
     Returns
     -------
-    dict
+    cclibSchema
         Dictionary of results from quacc.schemas.cclib.summarize_run
     """
     atoms = atoms if isinstance(atoms, Atoms) else atoms["atoms"]