+++ quacc/recipes/tblite/core.py
@@ -11,11 +11,16 @@ from ase.optimize import FIRE
 from monty.dev import requires
 
 from quacc.schemas.ase import (
+    OptSchema,
+    RunSchema,
+    ThermoSchema,
+    VibSchema,
     summarize_opt_run,
     summarize_run,
     summarize_thermo_run,
     summarize_vib_run,
 )
+from quacc.schemas.atoms import AtomsSchema
 from quacc.util.calc import run_ase_opt, run_ase_vib, run_calc
 from quacc.util.thermo import ideal_gas
 
@@ -28,10 +33,10 @@ except ImportError:
 @ct.electron
 @requires(TBLite, "tblite must be installed. Try pip install tblite[ase]")
 def static_job(
-    atoms: Atoms | dict[Literal["atoms"], Atoms],
+    atoms: Atoms | dict,
     method: Literal["GFN1-xTB", "GFN2-xTB", "IPEA1-xTB"] = "GFN2-xTB",
     calc_kwargs: dict | None = None,
-) -> dict:
+) -> RunSchema:
     """
     Carry out a single-point calculation.
 
@@ -46,7 +51,7 @@ def static_job(
 
     Returns
     -------
-    dict
+    RunSchema
         Dictionary of results from quacc.schemas.ase.summarize_run
     """
     atoms = atoms if isinstance(atoms, Atoms) else atoms["atoms"]
@@ -64,11 +69,11 @@ def static_job(
 @ct.electron
 @requires(TBLite, "tblite must be installed. Try pip install tblite[ase]")
 def relax_job(
-    atoms: Atoms | dict[Literal["atoms"], Atoms],
+    atoms: Atoms | dict,
     method: Literal["GFN1-xTB", "GFN2-xTB", "IPEA1-xTB"] = "GFN2-xTB",
     calc_kwargs: dict | None = None,
     opt_swaps: dict | None = None,
-) -> dict:
+) -> OptSchema:
     """
     Relax a structure.
 
@@ -86,7 +91,7 @@ def relax_job(
 
     Returns
     -------
-    dict
+    OptSchema
         Dictionary of results from quacc.schemas.ase.summarize_opt_run
     """
     atoms = atoms if isinstance(atoms, Atoms) else atoms["atoms"]
@@ -105,14 +110,14 @@ def relax_job(
 @ct.electron
 @requires(TBLite, "tblite must be installed. Try pip install tblite[ase]")
 def freq_job(
-    atoms: Atoms | dict[Literal["atoms"], Atoms],
+    atoms: Atoms | dict,
     method: Literal["GFN1-xTB", "GFN2-xTB", "IPEA1-xTB"] = "GFN2-xTB",
     energy: float = 0.0,
     temperature: float = 298.15,
     pressure: float = 1.0,
     calc_kwargs: dict | None = None,
     vib_kwargs: dict | None = None,
-) -> dict:
+) -> dict[Literal["vib", "thermo"], VibSchema | ThermoSchema]:
     """
     Run a frequency job and calculate thermochemistry.