+++ quacc/recipes/vasp/slabs.py
@@ -1,23 +1,23 @@
 """Recipes for slabs"""
 from __future__ import annotations
 
-from typing import Literal
+from typing import TYPE_CHECKING, Literal
 
 import covalent as ct
 from ase import Atoms
 
 from quacc.calculators.vasp import Vasp
-from quacc.schemas.vasp import summarize_run
+from quacc.schemas.vasp import VaspSchema, summarize_run
 from quacc.util.calc import run_calc
 from quacc.util.slabs import make_adsorbate_structures, make_max_slabs_from_bulk
 
 
 @ct.electron
 def slab_static_job(
-    atoms: Atoms | dict[Literal["atoms"], Atoms],
+    atoms: Atoms | dict,
     preset: str | None = None,
     calc_swaps: dict | None = None,
-) -> dict:
+) -> VaspSchema:
     """
     Function to carry out a single-point calculation on a slab.
 
@@ -42,7 +42,7 @@ def slab_static_job(
 
     Returns
     -------
-    dict
+    VaspSchema
         Dictionary of results from quacc.schemas.vasp.summarize_run
     """
     atoms = atoms if isinstance(atoms, Atoms) else atoms["atoms"]
@@ -69,10 +69,10 @@ def slab_static_job(
 
 @ct.electron
 def slab_relax_job(
-    atoms: Atoms | dict[Literal["atoms"], Atoms],
+    atoms: Atoms | dict,
     preset: str | None = None,
     calc_swaps: dict | None = None,
-) -> dict:
+) -> VaspSchema:
     """
     Function to relax a slab.
 
@@ -96,7 +96,7 @@ def slab_relax_job(
 
     Returns
     -------
-    dict
+    VaspSchema
         Dictionary of results from quacc.schemas.vasp.summarize_run
     """
     atoms = atoms if isinstance(atoms, Atoms) else atoms["atoms"]
@@ -128,7 +128,7 @@ def bulk_to_slabs_flow(
     slab_static_electron: ct.electron | None = slab_static_job,
     slab_relax_kwargs: dict | None = None,
     slab_static_kwargs: dict | None = None,
-) -> list[dict]:
+) -> list[VaspSchema]:
     """
     Workflow consisting of:
 
@@ -155,7 +155,7 @@ def bulk_to_slabs_flow(
 
     Returns
     -------
-    list[dict]
+    list[VaspSchema]
         List of dictionary results from quacc.schemas.vasp.summarize_run
     """
 
@@ -195,7 +195,7 @@ def slab_to_ads_flow(
     slab_static_electron: ct.electron | None = ct.electron(slab_static_job),
     slab_relax_kwargs: dict | None = None,
     slab_static_kwargs: dict | None = None,
-) -> dict:
+) -> list[VaspSchema]:
     """
     Workflow consisting of:
     1. Slab-adsorbate generation
@@ -218,6 +218,11 @@ def slab_to_ads_flow(
         Additional keyword arguments to pass to the relaxation calculation.
     slab_static_kwargs
         Additional keyword arguments to pass to the static calculation.
+
+    Returns
+    -------
+    list[VaspSchema]
+        List of dictionaries of results from quacc.schemas.vasp.summarize_run
     """
 
     slab_relax_kwargs = slab_relax_kwargs or {}